Deprecate old CLI console output formatters
A minor refactoring of some deprecated classes is included.
