+++ cli/src/main/java/bisq/cli/TransactionFormat.java
@@ -25,6 +25,7 @@ import static bisq.cli.ColumnHeaderConstants.*;
 import static bisq.cli.CurrencyFormat.formatSatoshis;
 import static com.google.common.base.Strings.padEnd;
 
+@Deprecated
 @VisibleForTesting
 public class TransactionFormat {
